+++ src/test/scala/cookbook/core/Serving_Static_Content_Example.scala
@@ -1,6 +1,5 @@
 package cookbook.core
 
-import io.fintrospect.ResourceLoader
 
 // fintrospect-core
 object Serving_Static_Content_Example extends App {
@@ -8,9 +7,9 @@ object Serving_Static_Content_Example extends App {
   import com.twitter.finagle.Http
   import com.twitter.finagle.http.path.Root
   import com.twitter.util.Await.ready
-  import io.fintrospect.StaticModule
+  import io.fintrospect.{Module, ResourceLoader, StaticModule}
 
-  val module: StaticModule = StaticModule(Root, ResourceLoader.Directory("."))
+  val module: Module = StaticModule(Root, ResourceLoader.Directory("."))
 
   ready(Http.serve(":9999", module.toService))
 }