+++ src/test/scala/presentation/_2/SearchApp.scala
@@ -14,7 +14,7 @@ import presentation.Books
 class SearchApp(books: Books) {
   def search() = Service.mk[Request, Response] { _ => Ok(books.titles().toString()) }
 
-  private val apiInfo = ApiInfo("search some books", "1.0", Option("an api for searching our book collection"))
+  private val apiInfo = ApiInfo("search some books", "1.0", "an api for searching our book collection")
 
   val service = RouteModule(Root, Swagger2dot0Json(apiInfo))
     .withRoute(RouteSpec("search books").at(Get) / "search" bindTo search)