+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 <h1 class="githubonly">Roadmap/Release Notes/Migration Guide</h1>
 
-The main API is very stable, but expect some amount of breaking changes around major releases as new features are added. All breaking changes are documented with a migration step where possible.
+The main API is stable, but expect some amount of breaking changes around major releases as new features are added. All breaking changes are documented with a migration step where possible.
 
 ## Backlog
 - Add strict `Content-Type` header checking for `RouteSpec`s based on `consuming()` and `withBody()` settings. Option for non-strictness.
@@ -11,6 +11,7 @@ The main API is very stable, but expect some amount of breaking changes around m
 - Added missing `number()` method for Double on JSON format objects
 - Bugfix #32 - `Circe` displaying ints as decimal when using `JsonFormat.number()`
 - (Small) breaking: Replaced function with `Service[T: Boolean]` in `ApiKey`
+- (Small) breaking: `ApiInfo` description field is now a nullable `String` instead of a nullable `Option[String]`. To fix, unwrap the option :)
 
 ## 14.0.0
 - Upgrade to various dependency versions, including `finagle-http 6.41.0`.